refactor: Make Table component more modular, #143
